Assume test assets are precompiled
Why?  First of all, live compiling of assets is true by default in the
test environment in rails so we need to get back to the defaults.

Sadly, this causes a 2 minute delay running any UI spec that calls
image_path/asset_path directly or indirectly.

Flamegraphs from rbspy show that the 2 minutes delay running a local
ui-classic test that calls image_path/asset_path is in sprocket-rails
in the call to asset_precompiled? which calls precompiled_assets(true)
if you're not caching classes, such as test.  It then needs to look
through all asset paths for precompiled assets to build a manifest and
ends up calling stat_tree / stat_directory in all these paths, such
as the node_modules.  We can assume the assets are precompiled as we
compile them on demand and can avoid this 2 minute stat'ing of files and
directories.  It's unclear if this has any negative consequences.

https://github.com/rails/sprockets-rails/blob/v3.2.1/lib/sprockets/railtie.rb#L39