Allow skipping initrd option in iPXE scripts
Previously, when we generated the MAC-specific iPXE script, we
unconditionally added the initrd command into it. Calling initrd
command without a parameter is not fatal, but it does pollute the
boot logs unnecessarily.

In this commit, we added simple if clause to the code, responsible for
generating the iPXE scripts, that leaves out the initrd command if the
PxeImage has initrd field set to nil.

Note that the code for generating the PXELINUX menus already had this
condition in place.