[configuration_script_source_spec] Avoid before(:all)
It was determined that using a `before(:all)` to setup the local "fake"
git repo might have file system contamination issues when trying to run
parallel specs.

Since this operation is fairly quick, moving it to a `before` with the
rest of the work there has little noticeable performance issues by
adding the FileIO calls for each spec.

Also, in addition to the constants being turned into `let` calls, they
also use `Dir.mktmpdir` to make sure there is a unique `clone_dir` for
each spec.