Add credential handling to Ansible::Runner
This creates a null-object-like class which will return a subclass
based on the type of credential id passed to it.

Each subclass will implement methods which will augment the environment
that ansible-runner executes in. Specifically they may return a hash
representing environment variables, extra vars, command line arguments
(to ansible-playbook) and they may write out the password and ssh key
files.

This commit implements a MachineCredential subclass as a start