[WIP][EmbeddedAnsible] Fix key unlock machine creds
`MachineCredential` records that have a `ssh_key_unlock` value aren't
respected in `ansible-runner` and will spin forever when running.

This adds the necessary value to the `write_password_file` to allow the
key unlock to happen in the `ssh-agent` calls when invoked via
`ansible-runner`.

* * *

Side Note:  The changes to `automation_manager/machine_credential.rb`
were because I was looking at the file and noticed the odd spacing.
Suprised `rubocop` didn't pick it up, but oh well.