[ansible_runner] Add VaultCredential
NOTE:  The `--vault-password-file` is used instead of populating the
`env/passwords` file here in case multiple credentials were used at the
same time.  This differs from how `awx` does things:

  https://github.com/ansible/awx/blob/1242ee2b/awx/main/tasks.py#L1554

Where as passwords expect file yaml file is used, but there is also
contextual awareness of all of the passwords being added at the time of
writing the file, where in the current case of MIQ, this is done for
each credential type, so we don't have the context available at when
writing the files.