[ansible_runner][MachineCredential] Clobber existing only
When creating a env/password file in MachineCredential, only inject new
values into the file if it already exists, but don't overwrite data that
wouldn't otherwise be re-written.

Example:  if an key for unlocking an encrypted SSH key exists, don't
delete that data while adding the other existing data.

A helper method, `.initialize_password_data` was added to the top level
`Ansible::Runner::Credential` class to share loading the existing data
or creating a new hash.