FirmwareRegistry create
It is currently cumbersome to add a new FirmwareRegistry
because it's split into three models:

- FirmwareRegistry
- Authentication
- Endpoint

Furthermore, each different type of registry can potentially
require different authentication/endpoint parameters.

With this commit we therefore introduce a utility method that
adds a new FirmwareRegistry, Authentication and Endpoint based
on options hash for us.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>