[MiqQueue] Add MiqQueue.broadcast
Adds a helper method to `MiqQueue` to enable broadcasting a job to all
servers.

Currently targets only all active servers, and is not filterable by
`:zone`/`:role` and an error will be raised if passed..  Furthermore,
those keys are set to `nil` so `MiqQueue.put` won't automatically apply
a filter to them when creating a job.