Cleanup the temporary checkout created by Embedded Ansible runs
This moves the knowledge of creation and destruction of the temporary
directory into the AnsibleRunnerWorkflow.  AnsibleRunnerWorkflow now
supports both a direct playbook_path (for on-disk playbooks) or a
configuration_script_source_id + playbook_relative_path pair of options.
If that pair is present, then the workflow knows it must checkout the
git repository contents first, and then later clean it up.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1741631