[EmbeddedAnsible] Add verbosity to job UI
When creating a new `EmbeddedAnsible::AutomationManager::Job` (aka, an
`OrchestrationStack`), make sure we are saving the verbosity from the
options we are running the job with.

We can fetch this after, but the service template (?) might change in
the future, so save off this value as part of the job record to ensure
we are keeping track of this value when it was run.