[EmbeddedAnsible] Add "job_plays" to job UI
This was previous functionality, though not really useful now we are
using runner as we only allow a single play per service.  That said,
this restores that functionality by providing an array of job plays.

The implementation here differs from what is provided in the ansible
tower provider:

https://github.com/ManageIQ/manageiq-providers-ansible_tower/blob/88aa135c/app/models/manageiq/providers/ansible_tower/shared/automation_manager/job.rb#L70-L72
https://github.com/ManageIQ/manageiq-providers-ansible_tower/blob/88aa135c/app/models/manageiq/providers/ansible_tower/shared/automation_manager/job.rb#L117-L144

Where a `OrchestrationStackResource` is created for each play instead of
using this temp Struct for working with the UI.

This might be desired if we are going to make use of this information in
other ways, but this solution was the quickest to implement at the time.