Add a tool to cleanup duplicate host guest_devices
As a result of the bug fixed by
[1691109](https://bugzilla.redhat.com/show_bug.cgi?id=1691109)
there are a lot of duplicate guest devices.  Now that the query is
correct EmsRefresh finds all of these and tries to delete them.  This
causes the RefreshWorker to use too much memory and get killed by the
MiqServer before it can clean them up.

This utility destroys them allowing the RefreshWorker to continue.