log_backtrace won't print backtrace for MiqException errors
See: https://github.com/ManageIQ/manageiq-gems-pending/blob/hammer/lib/gems/pending/util/vmdb-logger.rb#L106

This will still print as an ERROR level but the backtrace will not be printed to the log as this is a known
error condition where the backtrace is not needed.
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1729166

Note, embedded ansible was converted to use runner starting in ivanchuk, therefore this is a hammer only change.

See: https://github.com/ManageIQ/manageiq/pull/18687