[POC] Add MiqQueueLogFilterer for MiqQueue
Adds a helper class for MiqQueue that is in charge of filtering out
arguments that potentially contain sensitive from being logged by
MiqQueue.format_full_log_msg.

This currently requires hard-coding the methods that need to be filtered
on, but ideally this could be expanded on in the future to pluggable
instead of requiring core to be controlling this data.