ServiceOrder - add .v2v? virtual attribute
ordinarily, ServiceOrder is a set of orderable requests, bunched together in a shopping cart.
But v2v is also using ServiceOrder for In Progress migration plans.

Those ServiceOrders do not come from the shopping cart and do not contain regular provisioning requests.
They do contain requests of type ServiceTemplateTransformationPlanRequest.

We need to be able to tell these v2v service orders apart from the others,
because v2v migrations should not be displayed in the list of ordered services in SUI.

=> adding a virtual attribute so that we can use the API to filter out v2v service orders