Removing groups changes users current group
When removing a group that is a user's current_group_id, change it.

The {before,after}_destroy hooks work differently in rails 5.2

The user-group association is now removed before the cleanup.
So using this association to lookup users will always look at
no users.

Removing this dependency focuses on users who have a stale current_group.
Which is the basic premise of this task