Pin sprockets to ~>3.0 to fix missing manifest.js
With sprockets 4.0 we get the following error building the UI assets:
```
rails aborted!
Sprockets::Railtie::ManifestNeededError: Expected to find a manifest file in `app/assets/config/manifest.js`
But did not, please create this file and use it to link any assets that need
to be rendered by your app:
```

Pinning to ~>3.0 fixes this issue.