Base monolithic container image on pods ui worker
To do this I overwrote the entrypoint from the base image with
what is mostly the previous appliance initialize script.

The main changes I made were to add the server start at the end
and to remove the old container scripts references by pasting
the v2 key writing function into where it was previously called.

Additionally I removed starting memcached from the block that only
gets called if the database doesn't exist. We should start memcached
regardless.

This should allow the container to be started after a clean stop
Fixes #17274