Remove the unused InventoryCollectorWorker
The InventoryCollectorWorker was an experimental streaming refresh
worker for kubernetes/openshift but it put refresh targets on MiqQueue
rather than saving them in the same worker like the VMware
streaming refresh worker does.

If we revisit this in the future this will likely look more like the
VMware one where there is a thread in the RefreshWorker subscribed to
the watches and the main thread doing save_inventory.