Add MiqWorkerType
This adds a model which will replace the worker type constant.
It seeds the table based on the leaf subclasses of MiqWorker
and each worker class needs to define its own attributes, specifically
its priority in the kill order and bundler groups.

This allows new workers to be added by simply adding a new subclass
of MiqWorker rather than maintaining a list of worker classes in the
core repo.

Additionally this brings us closer to the ability to run a separate
process that will access the database and determine which workers
should be running on which servers, all outside of the main application
code-base.