metrics: move split_capture_intervals to provider
target is passed in since self is no longer the individual target.
This commit is probably the one that threw me the most in this chain,
because the kind_of will silently fail if you don't add the target there.

This code in general is VMWare specific, so getting it into provider
specific class is key.