Query for the worker count directly for concurrent smart proxy jobs
Previously this value would be saved to the server row by #update_capabilities
but that was likely a relic of a time when settings were only available
for the local server. Now we can query any server's settings from
a process on any other server so we can remove a cache.