Move the work a server does because config changes to #reload_settings
A call to this method is enqueued by a worker when it changes the
settings for a server. Previously, this method would reload the
server process's settings and this would set the last_changed
attribute on Vmdb::Settings. This attribute would later indicate to
the server to do some more work because its settings had changed
recently. This condition was checked and the work was done in
the #sync_monitor method.

This commit changes this flow so the work that was previously done
in #sync_monitor if the settings had been changed is now just done
directly when the server is told that it should reload its settings.

This simplifies the #sync_monitor method and also allows us to
remove Vmdb::Settings.last_loaded attribute entirely.