[GitWorktree] Only use non-bare clone when needed
In the previous commit, `.checkout_at` was added and `.checkout` was
redefined to use full `git-clone` actions to handle cases universally
for all repos.

This change updates that so this strategy is only used if
`Rugged::SubmoduleError` is raised, and we fallback to the new strategy.
Otherwise we favor the much lighter `.checkout_to` strategy, which
doesn't require a new clone of the repository to do.