[MiqReport] Handle duplicate Default MiqReports
This is a handler for "bad data" in the MiqReports table and allows for
an upgrade to happen while still retaining the data that exists.

This is not ideal in general, but seems like the best/safest solution to
avoid users getting an error that might not fully make sense.