Add an owner reference to the orchestrator pod for all created objects
This will allow anything we create to be automatically removed when
the orchestrator pod is removed for any reason. If the pod is cleanly
scaled down we will continue to tell the workers to exit cleanly, but
if the orchestrator is abruptly removed this will ensure everything
is cleaned up properly.

Fixes https://github.com/ManageIQ/manageiq-pods/issues/428