Use rails for exists clause
rails suggests model.where(id: model.joins(:other_model).where('a=b'))

we did the same. this method expects arel coming back, that is the
reason we used the arel nodes.

using primary_attribute.in(...) botched up the in clause, so we
resorted to the longer version of declaring the in clause.

wish it was not joining to Vm, but that is the ruby produced.