Add integration.rake tasks
Helper tasks that is used to run the necessary MiqWorker components for
running integration tests.

CYPRESS_DEV
-----------

Note on different modes:

- no `CYPRESS_DEV`:  "prod mode" where assets are static (CI, "faster")
- w/ `CYPRESS_DEV`:  dev mode where javascript assets might be edited

When using $CYPRESS_DEV, it is assumed that the user is modifying specs
or what they are testing, so being able to change those without having
to re-compile is ideal, but on CI, compiling should be a speed up for
the specs overall, and more "production like".