Move the systemd env var check into the systemd_worker? method
When starting a worker we were checking an env var and settings, when
stopping a worker we were just checking settings.  This meant that if
the env var wasn't set we would start the worker with spawn and try to
stop it with systemd which will fail because the unit doesn't exist.