Raname the key_pairs collection to match model name
The model class is called AuthKeyPair and by not having this match the
name of the collection it means that the auto_model_class doesn't work,
forcing all providers to manually set the model_class.

It is much easier to name the collection to match the model name and
just override the association because that doesn't change with provider
subclasses.