Handle system errors in workers
If a system error occurs (i.e. an Exception that isn't StandardError
derived), then we weren't handling that and the error would just bubble
out as an exception to STDOUT.  This has the strange side effect of
being logged, but without the logger prefix, thus appearing as part of
the previous line. If that line were an "INFO", then it would appear as
an "INFO" line.