[DONOTMERGE][WIP] Add ManageIQ::Integration::DatabaseCleaner
This is a custom cleaner which as part of the last phase of seeding (for
integration ENV only), will create a "template" DB that can be used to
restore the database to it's previous state.

While a `DROP` then `CREATE` using the template would be nice and more
in line with the standard use cases of template DBs, this isn't really
feasible while keeping the application open (and thus, connections to
the database intact).  So the "template" is just a copy of the database
base "post seed".

To "clean", we identify tables that "post seed" have data in them,
truncate all but "skippable" tables (miq_servers, miq_workers... aka
"what keep the application running")), and then reseed them using a
Foreign Data Wrapper for postgres that points the template DB.

This class helps set that up as part of seeding, and the info that is
collected as part of the setup phase is then cached (in the
`seed_config.yml`) and used by what ever process will need to do the
"clean" operation.