Don't disconnect resource pools in refresh_new_target
When saving a new VM using the legacy refresh_new_target we rely on not
deleting records that aren't related to the new target for this to work.

The save_resource_pools_inventory method was incorrectly assuming that
not setting target = ems and then later checking target == ems
(aka a full refresh) to see what it should delete, but in the case
of refresh_new_target the target is already the ems so every
resource pool except the one the VM was on was being deleted.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1846273