Remove EmsRefresh.refresh_new_target
EmsRefresh.refresh_new_target was how we did targeted refresh of a new
target (e.g. something that wasn't in the database yet) prior to the
addition of graph refresh and the InventoryRefresh::Target concept.

RHV is the last consumer of this method for their event handling and
once it is removed we can delete this method and the code supporting it.