[FIXUP][integration.rake] Reset `environments` path
When doing `rake :env`, the `config.paths` method has already been
called/memoized when the `config/application.rb`, so that causes the
following to be called:

    paths.add "config/environments", glob: "#{Rails.env}.rb"

Which even before `:env` has a chance to run, it will set the `:glob`
attribute to the "default `Rails.env`", which in turn only allows
loading the include `developement.rb` when the initializer to load the
environment config for `Vmdb::Applciation` is finally trigger.

While not the best of solutions, I prefer it to having to have super
long commands like this:

    $ RAILS_ENV=integration rake integration:seed:db

Where `integration` is implied by the task itself (and helps "codify"
workflows).