[AuthenticationMixin] Fix .authentication_status
As mentioned in the **lengthy** comment, this ensures that when using
`:authentication_status_severity_level` there is the following stanza in
the WHERE clause:

    '"authentications"."relationship_type" = ?'

Specifically, when `.authentication_status` is used in the `.select` of
a `ActiveRecord` query.

Previously, it would generate something like this (using Host as the
example class here):

    SELECT "hosts".*,
           (SELECT  "authentications"."status"
            FROM "authentications"
            WHERE "authentications"."resource_id" = "hosts"."id"
            ORDER BY CASE
                       WHEN LOWER("authentications"."status") = ''            THEN -1
                       WHEN LOWER("authentications"."status") = 'valid'       THEN 0
                       WHEN LOWER("authentications"."status") = 'none'        THEN 1
                       WHEN LOWER("authentications"."status") = 'incomplete'  THEN 1
                       WHEN LOWER("authentications"."status") = 'error'       THEN 2
                       WHEN LOWER("authentications"."status") = 'unreachable' THEN 2
                       WHEN LOWER("authentications"."status") = 'invalid'     THEN 3
                       ELSE -1
                     END DESC
            LIMIT 1) AS "authentication_status"
    FROM "hosts"
    WHERE "hosts"."type" IN ('ManageIQ::Providers::Redhat::InfraManager::Host')
      AND (hosts.id IN (2))

Where you will see the nested SELECT matches on `resource_id`, but not
`resource_type`.  In this case, if there was a EMS with the same id as
the host in question, the EMS's `Authentication` record could be
returned instead of what is associated with the Host.  In particular, if
the Host happened to have 0 Authentication records, it means that when
doing this particular query, it could appear to be given a status of
"None" in one case (the default when there are no records), and a say
"Valid" state when the query is used via a nested SELECT and EMS's
Authentication record is "Valid".

* * *

The only side effect of this approach (aside from the ugly code it
produces) is that there are now cases where the `resource_type` check
happens twice within the same query.  Since they results should be
equivalent, this shouldn't matter besides slightly increasing the query
string length sent to Postgres.