Improve thread-safety by avoiding a write based on a read
There is no atomic get/set operation on a simple concurrent hash so we can avoid
doing a write based on a read through a possible two thread safe writes.  This
double write of the same thing only happens when first saving a pod during the
initial get_pods or from an update from watch_pods.