[ManageIQ::Session] Add .store
Adds a quick way of accessing the session store from outside of the
Rails session middleware.  Allows our backend to more easily interact
with the SessionStore via the ManageIQ::Session interface.

The `.fetch_store_from_rails` is lazily fetched to ensure that the
middleware stack is built before we try and access it.  Also, the code
for that is pretty gnarly... blame Jason...