[ManageIQ::Session] Patch .delete_sessions
Adds/Patches a .delete_sessions method to all of the session store
adapters that can delete a collection of ids that are passed to it.

Implements a `FakeRequest` object that can be passed as part of the args
for `delete_sessions` for each store that can act like a
`ActionDispatch::Request` object, but simply provides the `.env` var as
an empty hash.  This allows `.delete_session` to be used with it's
original call signatures (for the adapters that require it) without
having to dive deep into the internals of each adapter to call custom
methods that do the same thing.

`:drop` is basically assumed for all adapters since there is no good way
to keep track of the existing data for the session when dealing with
multiple sessions.