configure db query matchers for all repos
for performance counters, we're using db_query_matchers.
Without configuration, it also counts schema queries.
Which is typically not a problem, but it does add some sporadic
counts.

this change filters them out so we can focus on the model queries.
Which are what mostly concern us since the schema queries are
only run once