consistent child manager saves
most providers have child providers. They all seem to
propagate attributes to the child with a different set of attributes.
They also can not be initiated in maintenance mode because not all
attributes are copied across.

This gives a common method for providers to consistently propagate
the attributes to the child managers

we were able to drop the cinder and swift helpers because the are no longer
referenced from openstack