[rails6][EvmApplication] Update MigrationContext.new
This commit in Rails:

https://github.com/rails/rails/commit/7cc27d749c3563e6b278ad01d233cb92ea3b7935

Updated the ActiveRecord::MigrationContext.new arguments to also include
a second argument that passed the base class(?) for all migrations.

Originally, this commit was coded to basically use the default value
from the `ActiveRecord::Migration#copy` method:

https://github.com/rails/rails/blob/7cc27d749c3563e6b278ad01d233cb92ea3b7935/activerecord/lib/active_record/migration.rb#L887

    schema_migration = options[:schema_migration] || ActiveRecord::SchemaMigration

Which seemed to cause trouble when running the specs.  Using the author
of PaperTrail's comment as inspiration:

https://github.com/rails/rails/pull/36439#issuecomment-517379182

Making use of ::ActiveRecord::Base.connection.schema_migration seems to
work as a good second arg for our use cases.