[EmbeddedAnsible] Ensure newline for :ssh_key_data
SSH formats like `OPENSSH` require that a newline exist on the last
line, otherwise it is considered an invalid format.

This adds a `before_validation` callback to the model to ensure that it
adds a newline to the key (if a key exists) in case it was stripped off
by the UI or via other means.