Eager load api collection classes in ui workers
Similar change but now for the UI worker:
https://github.com/ManageIQ/manageiq/pull/20072

"Allowing threads to autoload code requires us to spend time tracking down bugs
in code loading and insert disgusting rails interlock blocks to "protect"
various places that could load constants in threads.

This solution tries to avoid these workarounds by eager loading all of the
collection classes and their descendants before any requests are made, since
we're likely to need to load them anyway."

Note, the api classes from the manageiq-api repo is included in the manageiq_default
gemset, so it's safe to use it.

Note, web console worker is another web server worker, so we're only doing UI here
in addition to the web service worker we did previously, but we're not doing it for
all web server workers.