replace cattr_accessor with class attribute
i'd rather this were a class attribute

same change as was made in https://github.com/ManageIQ/manageiq/pull/20795, with probably the same level of concern — see https://github.com/ManageIQ/manageiq/pull/20795#issuecomment-726356151. this change is unlikely to matter but in principle if we don't need to use this I still think we shouldn't.