Add basic script to print pod throttling information
We could possibly throw this into a bash script that all images have
and therefore we could run within the httpd/postgresql pods too.

Either this, this is just a first cut, first try to make sure it works.
We can change this to a ruby method/shell function that outputs % throttled
or whatever we want.  We can then choose to log this or raise events based on it.

This is what it looks like when I run this using images containing this script:

```
% for I in `oc get pods -o wide | grep -v NAME | awk '{ print $1 }'`; do echo $I; oc exec -t $I  ruby /var/www/miq/vmdb/tools/pod_throttling.rb; done

1-event-handler-86bc8c48-j6527
0.0022497187851518562, nr_throttled/nr_periods, 4/1778.0

1-generic-76c4cc89df-j8zk8
0.015989036089538604, nr_throttled/nr_periods, 35/2189.0

1-priority-6db6fb54bc-s6fvk
0.011067475901463763, nr_throttled/nr_periods, 31/2801.0

...
1-ui-64bb7d5fb8-f28nw
0.006785411365564037, nr_throttled/nr_periods, 8/1179.0
1-web-service-bff886b46-4cwbr
0.001834862385321101, nr_throttled/nr_periods, 3/1635.0
...
orchestrator-dd875c564-6lvvq
No periods monitored: no limit set?
...

```