Run MiqServer.status_update in server process
Fixes #20835
[Errno::ESRCH]: No such process ... when run from a worker.

In appliance/systemd, you can run status_update in the server's workers.
In podified, each process is isolated so we must run this, specifically
process_status, in the server process.

It didn't make sense to run this is in the worker processes anyway so
change all installations to run this in the server process.