Fix simulate_queue_worker not marking queue records as state dequeue
When a queue_worker performs a `MiqQueue.get()` the messages are marked
with `msg.update!(:state => STATE_DEQUEUE` before being returned to the
worker to be operated on.

This is critical if the queue item being executed does a
`MiqQueue.put_or_update` which might update the current queue item.  A
good example of this is `EmsRefresh.refresh` which queues refreshes for
its child managers.

If the current queue item being run isn't marked as `"dequeue"` then
instead of creating a new queue item for the child managers, the child
targets are merged into the current queue item which is then deleted
after it is completed.  This results in the child managers never getting
refreshed.