hide ancestry from the api
Ancestry is an encoded field with the parent id.
This is only for internal consumption.

If someone asks why this was removed, we can provide alternative options
for discovering the ancestry.

I had a customer say that this field was broken because it didn't contain
something - but they weren't using the field correctly. That is why I went
on this hidden attribute journey in the first place. I wanted to limit
our exposure to customers using internal fields incorrectly.