[evm_database_ops_spec.rb] Stub proper method
This was changed as part of an effort in `manageiq-gems-pending` to
remove `MiqUtil.runcmd`:

https://github.com/ManageIQ/manageiq-gems-pending//commit/729d5c00ac877a12aca72ac2efff3beda8eda9cd

And as part of that, the method `runcmd_with_logging` was changed to
`run_command_with_logging` in the `PostgresAdmin` class.

While an alias exists for it, the internal method inside of the class
was changed, and this spec stubs the internals of that class (in
otherwords... not the best test in the world...).

Simply stubbing the proper method allows the tests to pass, and making
the spec more resilient can be left for another day.  #punt