Fix sporadic test failure when MiqEvent is not yet defined
The test is checking that MiqEvents were raised to the queue when space
is not available for the backup/dump feature, however depending on test
loading order, it's possible that MiqEvent is not defined.  The code
that raises the event has an explicit check for `defined?(::MiqEvent)`
because it can run outside of the Rails environment context from a rake
task.

This commit changes the test to ensure that the MiqEvent class is
defined, so we can verify that the event is raised.

Closes #20986