Prefer mocking invalid domains over VCR
spec/models/firmware_registry/rest_api_depot_spec.rb:36 was using a
vcr cassette, but that cassette was never committed.  As such, when
this spec runs it a) reaches out to DNS and b) in the case of home
users where the ISP intercepts invalid domains (e.g. https://searchassist.verizon.com)
it will create an untracked VCR file locally.

VCR will not write a cassette for invalid domains, since it raises a
SocketError.  So, this commit uses webmock to stub the request and
force a SocketError, which is what a normal, non-intercepted, invalid
domain request will look like.

Additionally, this commit changes from bad.host to example.invalid,
which is a recognized "invalid domain" as per RFC2606.