[AuthenticationMixin] Allow for has_one
Allows models to make use of the `AuthenticationMixin`, but have only a
single authentication relationship where appropriate.  It also requires
the model in question to implement a `set_authentication_type` class
method that will set `:type` and `:resource_type` for the given STI
authentication when including this class.

In the case of `Zone`, we only ever have 1 instance of the
`AuthUseridPassword` with a `type` of `:windows_domain`, so this makes
more sense for that to be used there.