Change spec to use Tempfile with a block
This spec has a sporadic test failure with the way it creates a tempfile.
Prior to this commit, the tempfile is created and closed before the spec
runs. However, if the garbarge collector runs after this, then Ruby will
delete the temp file.  This commit changes it to use the Tempfile.create
method with a block, which will prevent Ruby from deleting during the
test.