Run workers without (root) login shell option
We currently run a login shell what will `source`/load the evm environment.

Instead, this directly loads the environment directly and does not rely upon a bash login shell.

We are making this change to get away from relying upon root's bashrc configuration and
eventually run these are a non-root user.

I would have liked to have used `EnvironmentFile` instead of sourcing the script
but that change was non-trivial

This is a continuation of https://github.com/ManageIQ/manageiq/pull/20983

I will add `User=` and make all of these in one sweeping change to the other 22 repos.