Allow higher memory usage since exceeding it in podified is unforgiving
Appliance memory limits are treated with graceful restarts of worker processes
when they complete their work.  Conversely, podified is unforgiving in that tripped
memory limits result in an OOMKilled process in the pod with exit code 137, no
cleanup including a long running task left in limbo, and a restarted pod.

The downside of increasing memory limits in an already graceful appliance model
is less than the downside of pods getting OOMKilled, possibly repeatedly,
and leaving tasks incomplete.