Stop logging to files in production deployments
In our production deployments, we don't need the log files, and, in fact, they
create issues. In appliances, we now use journald and the logs are duplicated
in both /var/www/miq/vmdb/log as well as /var/log/messages (see also #21211).
In containers, the logs fill up the local volume unnecessarily as they are
also output to STDOUT, and STDOUT is preferable because it can be scraped by
a log aggregator. As such, this PR stops logging to the log files in
production.

Resolves #21004, #21211

Co-authored-by: Keenan Brock <keenan@thebrocks.net>