[.simulate_queue_worker] Honor :deliver_on
Previously, queue items that had `:deliver_on` in them were executed
right away by `simulate_queue_worker`, since the only thing it filters
by is `queue_name`.

This is problematic for things like importing zip files for automate
domains as it will upload file, and then queue it for deletion a day
later:

  https://github.com/ManageIQ/manageiq-ui-classic/blob/7b0bda1/app/services/automate_import_service.rb#L44-L50
  https://github.com/ManageIQ/manageiq-ui-classic/blob/7b0bda1/app/services/automate_import_service.rb#L94

Only when doing this with `simulate_queue_worker`, it deletes it right
away, and in many cases doesn't give it a chance to get processed before
it is deleted.

Honoring the `deliver_on` attribute fixes that issue.