[FIXUP][MiqSet] Handle member relationships
As a way to allow the `members` relationship, and "alias" forms of it,
the `inherited` method is being used to allow subclasses to define these
methods specifically.  This is basically required as it the `has_many`
requires a `:source` and `:source_type`, and those need to be targeted
at the particular relationship in question.

For some strange reason, `alias` was also not working with these, so to
solve this, instead of doing an `alias`, the different forms of the
method were all defined using a `.each`.

Also, abstract_class didn't appropriate here as it wouldn't allow
`MiqSet.all` to work, which seems like it might be useful, even if only
for dev.