[Tagging] Fix tagging with MiqSet
MiqSet, and associated models, assume that the :taggable_type will be
the subclass of whatever `MiqSet` is being used.  The default behavior
of `ActiveRecord` when dealing with a polymorphic association is to set
the value of the `_type` column to `polymorphic_name`:

    module ActiveRecord
      module Associations
        class BelongsToPolymorphicAssociation < BelongsToAssociation #:nodoc:

          private
            def replace_keys(record)
              super
              owner[reflection.foreign_type] = record ? record.class.polymorphic_name : nil
            end

Which by default for that is `base_class.name`:

    module ActiveRecord
      module Inheritance

        module ClassMethods

          def polymorphic_name
            base_class.name
          end

However, in the previous form of `acts_as_miq_set`, it wasn't
truly subclassed as it is now, so it would be saved as `subclass.name`.
A lot of code relies on this behavior, many other classes follow this
properly, and it would probably require a migration to fix old records
if we were to change it now.  So the option chosen here is to flip to
saving the `taggable_type` as `subclass.name` when the recorded being
associated with the tag is a `.kind_of? MiqSet`, otherwise do the
default.

This preserves with previous behavior without mucking with the cases
where this is done normally.