[Exports::CustomButtons] Refactor/Simplify
The previous implementation of this exporter used very abstract logic to
export the specific data items needed.  While incredibly DRY as a
result, it is also very hard to parse and when the underlying DB
structure changes, it is even harder to fix (#foreshadowing)

As a result, the change specifically makes it so the class is much more
direct in what it is exporting, and hopefully makes it easier to digest
on how it is working as well.