Set the `:manager` when building a TargetCollection
The `InventoryRefresh::TargetCollection` has a `:manager` and
`:manager_id` property.  We were only setting the `:manager_id` even
though we had the `ems` already looked up.  This saves us a query later
and makes it easier to use the EMS class name for building class names
based on target type.