Add proper expectation stubs for slow MiqRegion.destroy_region
We need to do proper expectations to ensure it's called the right number of times
with the correct arguments in some situations but we should not actually call
this in test as it walks all replicated tables and removes data for the region.

Locally, tests go from 31 to 0.7 seconds.

Before:
% be rspec spec/models/pglogical_subscription_spec.rb | grep Finished
Finished in 31.15 seconds (files took 11.66 seconds to load)

After:
% be rspec spec/models/pglogical_subscription_spec.rb | grep Finished
Finished in 0.7697 seconds (files took 11.78 seconds to load)