Create log dir if it doesn't already exist
The log dir is intentionally not part of the repository since we want to
use a separate volume mount for that in the appliances.

However, during release tagging, which uses simple rake tasks, creation
of the logger fails because the directory doesn't exist. This commit
creates the log dir on demand if it doesn't already exist.