Remove EvmDatabaseOps and related rake tasks
Well... mostly.

A single (set of) method(s) is needed as the appliance_console uses this
method to check the current database connection.

This code could be moved into the console itself, but since it was
released recently, this allows that code to still function until a new
version not requiring it can be released.