Use default HOME
we were creating an override for the HOME value for each
process created.

HOME=/root does not work for non root users
Also, it is easier to just set the variable in the systemd service
file rather than create a custom override.
Or set the value in /etc/defaults if it is universal