Add debugging for Ansible::Runner
Allows keeping the file system generated by the `Ansible::Runner` lib
(aka: `EmbeddedAnsible`) via two parameters:

- Setting `verbosity` to 5 (or greater)
- Setting `ANSIBLE_KEEP_REMOTE_FILES` as an ENV var to '1'

This will cause `cleanup_filesystem!` to no-op and allows keeping the
filesystem in place without requiring a code change and application
restart, and doesn't require new field on services.

Note: Greater than 5 for verbosity is invalid as far as Ansible is
concerned, but that is `.chomped(1,5)` in a different part of the code
base, so using "6" could be a way to say "I would like `-vvvvv` plus
extra debugging from `MIQ`".