Support both dalli 3.1+ and 3.0
* Remove the with_lock/with_block monkey patch

We originally added the with_lock wrapper to ensure the dalli lock and the active support dependencies lock played nice.

Our monkey patch added this here:
https://github.com/ManageIQ/manageiq/pull/7334

The AS::Dependencies lock had to play nice with the mutex found in dalli here:
https://github.com/petergoldstein/dalli/blob/v2.7.6/lib/rack/session/dalli.rb#L73-L74

Therefore we monkey patched dalli session to ensure this mutex would allow
AS::Dependencies lock to be released properly and avoid a deadlock betweeen the two
locks.

Dalli renamed with_lock to with_block in 2.7.7 (yes, a patch release):
https://github.com/petergoldstein/dalli/commit/9f9c508afab263a2451f2209c4396daf98d33a1b#diff-4b41e4b33e4b095e350045d4e727d2e889890cfc0610d434459f30a267423065

The lock in with_block was removed in dalli 3.0.0: https://github.com/petergoldstein/dalli/commit/7be38027f36cfd5e3e8dd9fe0d2d5843bf7292da

All that for this:
The lock is gone as of 3.0.0 so we don't need to wrap with_block anymore!

* Dalli 3.1.0 renamed session method destroy_session to delete_session.  To
avoid dynamic code, we're using simple conditionals to call the correct method
based on the version.

* Dalli 3.1.0's change to use Abstract::PersistedStore instead of
Abstract::Persisted changed the session id from a string to a SessionID object that can't
be coerced in finders.  We need to handle this for dalli 3.0/3.1 compatibility.
See: https://github.com/petergoldstein/dalli/commit/3abf71533c5288dd49c4756fb1c9f2989e263e89

* We now allow dalli both 3.0 and 3.1+.  We can require 3.1+ and eliminate some
conditionals once we're satisfied that 3.1 is working for us.