fix default_value_for(alias)
Default value is only working for actual columns in rails 6.1
and not for column aliases

All other provision workflows use source. This is the only one
left that is using `src_vm_id`

Switching over to using just `source_id` works but then the
`source_type` was not set, so this is defaulting both

A little concerned what would happen if we pass a non_vm for source_id
bit it looks like that is the only object that will be passed in.
And if `source_id` is passed in, they will also pass in the `source_type`

use default value with an actual column

     ActiveModel::MissingAttributeError:
       can't write unknown attribute `src_vm_id`