Allow updating gems before locking
Normally a bundle update would be sufficient for updating gems, but
because of the way we need to link in the appliance console, and avoid
bundler.d items, it's easier to just tack this into the
generate_lockfile method. In the future we might want to break this up
into two separate parts and also share some code with the rake release
command that also does something similar.