Rescue errors patching the deployments, log it, and skip this worker
This allows us to continue patching deployments for all worker's classes and
not unwind the stack to the rescue in the monitor method.

Previously, in monitor_workers:

```
    sync_from_system # An uncaught exception is raised

    sync_monitor # Skipped

    sync_workers # Skipped

    MiqWorker.status_update_all # Skipped

    cleanup_failed_workers # Skipped
```

This exception would be caught in monitor, meaning, we'd never run the other
methods after sync_from_system as long as this Kubeclient::ResourceNotFoundError
was being raised.  This was "fixed" as soon as the orphaned worker row, a row
without a pod (due to the deleted deployment), was cleaned up, which in this
case occured about 10 minutes later.

We should never blow up the stack all the way to monitor.