define virtual attributes without block
We do not want introspection a class at class load time.
So we are not able to derive a column type at load time.
We pass a block for the attribute type to allow us to derive the type at runtime.

Now that we no longer use this feature in virtual delegation, it does not
not make sense to keep it around just for deprecated attributes

We would like to use the rails attributes api in virtual attributes, and
that does not support this feature. So removing this feature makes that
transition easier.