Fix Ansible::Runner timeout waiting for pid
When we start the ansible-runner command we wait for command to start
before returning the async results object.

This had been waiting for a `pid` file in the top of the temp directory
to be created but in more recent versions that no longer is created.

We can instead wait for the `artifacts/result/command` file which is
created early in the run.