Prematurely stopping evmserverd is unnecessary
Required for https://github.com/ManageIQ/manageiq-postgres_ha_admin/pull/36

We restart evmserverd upon successful failover anyway, so it's pointless to stop
beforehand.  Let's be clear, failover in this code means we updated the
database.yml.  The logical replication handler doesn't even do anything in the
"failover" step.

Additionally, the before_failover hook cannot know if failover will be
successful, so it's best to only put actions here that are truly prerequisites
to failover.  Any actions we add to before_failover will need to be understood
and dealt with in the event of success or failover in the after hook.

Less is more.