Disable syndication by default
Previously we defaulted to true for syndication of events, inventory,
and metrics but we would only publish if miq_messaging/kafka was
configured, the expectation being if you went through the effort of
setting up kafka it was for the explicit purpose of syndication.

Now that we use kafka internally this assumption is no longer true and
this can lead to a lot of "wasted" traffic on kafka.