Extract active role syncing out of worker management
Part of the worker_management sync_monitor loop is to sync_active_roles
for the server, then notify workers of the change.

This seems like the wrong place for this since we have a RoleManagement
module already, and most of the worker_management code has to call up to
RoleManagement which to me indicates that it is in the wrong spot to
begin with.