metrics perf_capture: send batches
instead of sending one message per vm, send a batch of messages.
This is done by sending a list of target_ids to perf_capture

The batch size (or if batching is turned on at all) is configured in
the per ems settings.

before
======

- one queue message per target per interval per time range
- once all hosts collected, MiqTask synchronized and triggers a message to
  rollup hosts into clusters (this is purpose of the cb callback)
- queue helped us prevent duplicate collection requests

after
=====

- one queue message for multiple targets per interval per time range
- all hosts in a cluster are in a single message, so a rollup parameter
  triggers a message to rollup host into clusters

Since our total number of queue entries is in the dozens, we no longer need to
coordinate having multiple queue records per host. So the put_or_update can go away

performance
===========

|        ms |        bytes |     objects |query |   qry ms |     rows |`comments`
|       ---:|          ---:|         ---:|  ---:|      ---:|      ---:| ---
|  97,673.9 | 597,353,117* | 180,998,775 |3,280 | 37,923.2 |      280 |`capture-before`
|  92,179.5 | 677,534,487* | 180,999,121 |3,246 | 32,189.4 |      262 |`capture-after`
|           |              |             |  1%  |          |       6% | delta

The real savings is avoiding locks into the MiqTask table and dropping the put_or_update