remove ems and zone lookup for every rollup
We run rollups for a single ems, and a single zone.
But we were looking up an ems and my_zone for every resource rollup

optimization: pass the zone to miq, since it is known and we already
have the ems / zone in memory. We have had to do this optimization for a number
of other miq_queue calls, so it makes sense.

This code is using the parent object for the ems. Any one would do.
If this method gets converted to a static method,
then use any one of the resources' ems.

numbers
-------

Running capture for only 16 Vms. They generate:

ManageIQ::Providers::Vmware::InfraManager::HostEsx#perf_capture_realtime => 6
ManageIQ::Providers::Vmware::InfraManager::Vm#perf_capture_realtime => 16

So that will save 22 ems and 22 zone lookups (total number of vms + hosts)
Minor savings. But removes queries
But it clears up the number of queries and makes it easier to

|        ms |        bytes |    objects |query |   qry ms |     rows |`comments`
|       ---:|          ---:|        ---:|  ---:|      ---:|      ---:|  ---
|  19,312.2 | 138,947,654* | 32,019,635 |1,151 |  6,993.9 |      262 |`capture-vc91-250-before`
|  19,192.5 | 108,353,470* | 31,971,845 |1,117 |  6,869.5 |      228 |`capture-vc91-250-after`
|           |              |            |   3% |          |      13% | delta