Add logging when filtering alerts
Previously, we'd quietly filter alerts that weren't assigned, not enabled,
and not matching the desired event name.  Additionally, we'd not report when we
queued no alerts due to all alerts being filtered out.

This change avoids making logical changes to some "interesting" code but allows
us to better understand when alerts aren't not firing.