Call after_update_* from edit_with_params
When changes are made to the endpoints or authentications workers which
run with an open connection like event catchers and streaming refresh
workers have to be restarted.

This was done via `after_update_authentication` called from
`update_authentication`.  The issue is that this method is no longer
called when providers are updated via the API with DDF parameters.