Reduce filtering for pagination in Rbac.search
Before:

Some filtering can only be done in ruby.
For these pages, we download all records and run the filtering locally.
We were filtering every record and then limiting the records for display
This was generating a lot of extra work since we only need to filter the
records that can end up on the page.

After:

In the case that we are returning counts, it is not possible to avoid
filtering every record, but in the typical case (i.e.: `skip_counts = true`),
we are able to only filter the records up until the end of the appropriate page.
So we introduced filtering with a lazy enumerator. In the count case, we use
to_a to undo the lazy enumeration - this avoids running the lazy enumeration
multiple times.

We've just added yet another type that the targets variable can be.
There is the potential that not every type will be in the test cases and
a future change doesn't take this into account. I tried to add come comments
to show where I felt the issues could arise.

We are down to one caller that does not use skip_counts.
This is the main view page in the ui so it is popular by call volume
Removing this would remove a number of touch points in this search code.