Inject into bundler.d for a generated plugin
When generating a plugin or provider we inject the gem dependency into
the core `Gemfile` automatically for the user, but we do not add any
reference to the local plugin to the `bundler.d` overrides, and since no
generated plugin exists in `github.com:ManageIQ/` yet running `rails c`
and `bundle install` always fails until the developer adds the new
plugin to their overrides.

This change will detect if an overrides file already exists otherwise it
will create one.