Enhance factory for host with authentication(s)
I had introduced :with_authentication like ems and provider had,
but realized that authtype was more explicit. That can be introduced later
if we feel it is easier to read.

I'm reworking the way hosts are validating authentications and need a little more flexibility

I had wanted to change the ems factory to take these changes, also setting the resource in the authentication,
but this created MiqQueue entries and that caused all sorts of havoc in tests. I avoided these changes
so this PR did not modify too many other tests.