Rbac#filter prunes ruby in miq expression
rbac#filter/search can use an miq expression to filter the results.
These expressions can be ruby only, sql, or a combination of both.

The issue is the code was smart enough to send the sql the WHERE clause
that was sql friendly.
But when we were filtering in ruby, we performed all the sql filtering
and ruby filtering.

For 1k records, this takes a hit.

After

We send the sql only filters to sql and the ruby only filters to ruby.
In cases where the OR mixes sql friendly and ruby friendly parameters,
we still perform the sql friendly filters in ruby.
(logically it is the only way to do it)