Reduce filtering for pagination in Rbac.search
Overview
========

We use miq expressions to filter records.
We also perform pagination.

So for our example, we have 100 records, 5 records per page,
and we're showing the second page (skipping the first 5 records).

```
miq_e = -> (x) { print '.' ; x % 2 == 0 }
(1..100).select { |i| miq_e.call(i) }.drop(5).take(5)
Problem: It takes time to execute the miq expressions, and we are executing it once per record.
```

Goal: Reduce the number of executions of the miq expression.

Now lets be clear, the best optimization we can perform is converting the expression into something
sql friendly. But this PR assumes that converting into SQL is not possible.

Optimization theory time
========================

Lets only execute the query enough times to fill the current page.

If we show 5 records per page, on page 2, we skip the first 5 records, but show the next 5 records.
So we'll need to process 10 successful records until we have enough to fill the page.
If the miq expression marks every other record for display, then after processing 20 records,
we'll have 10 successful records. drop the first 5 and we have 5 for display.

Goal: Only execute the miq expressions 20 times, even with 100 records.

Before
======

We'll print a period for every miq expression run.

```
miq_e = -> (x) { print '.' ; x % 2 == 0 }
b=(1..100).select { |i| miq_e.call(i) }.drop(5).take(5).collect(&:to_i)
[b, b.size].inspect

.................................................................................................... (100)
```

We run the miq expression for every record (100 times).
But once that is done, we don't run it ever again.

After
=====

The lazy converts the range into a lazy enumerator. Future blocks detect this as a lazy enumerator and are only run just in time.
The drop does ask for records, as does the take. But the take stops asking for records well before the full 100 records are requested.
The collect really doesn't do anything, but I kept it in there to be closer to the actual result.
The to_a forces the full execution and converts the enumerator back into an array. An array is not lazy.

lazy came about because python seemed to use infinite ranges ( 1.. ) to do stuff, and ruby said we want something like that

```
miq_e = -> (x) { print '.' ; x % 2 == 0 }
a=(1..100).lazy.select { |i| miq_e.call(i) }.drop(5).take(5).collect(&:to_i)

[a, a.size].inspect

[a.to_a].inspect
....................# => "[[12, 14, 16, 18, 20]]"

[a.to_a, a.to_a.size].inspect
........................................# => "[[12, 14, 16, 18, 20], 5]"
```

- We need to use `to_a` to get the data out. Without it it doesn't even run the miq expression once.
- All blocks are converted to the lazy form down the line.
- We execute the miq expression the least number of times necessary.
- Every time the data accessed, we execute the miq expression.

best case scenario:

- We use `lazy` to only filter the rows when they are read.
- We use `to_a` to filter all records if we are counting. This avoids running the enumeration/filtering twice.
- We use `to_a` to convert any lazy enumeration to an array, so the external interface does not change.
- We use `to_a` after `drop()/take()` so the minimum number of records are processed.
- We do not use `lazy` or `to_a` for a query that is pure sql

When does this optimization NOT apply?
===

1. The ui-classic list view displays the total count (auth_count) [ref]
We still need to filter every record to get the total count, so this can not be used.
No `skip_count` means no optimization.

2. If we are displaying every record, then we need to filter every record.
No pagination means no optimization.

3. If we have no MiqExpression, then we are not filtering by miq expression
and don't need to optimize it.
No miq_expression means no optimization b/c it is already fast.

4. If the MiqExpression is only using sql columns, then the filtering and
pagination happens in the database. So we are only downloading the correct
number of records and we don't even need to filter in ruby. Typically these
return a scope anyway.
No ruby miq_expression means no optimization b/c it is already fast.

So... the API is a good example of a query that can leverage this optimization.

[ref]: (https://github.com/ManageIQ/manageiq-ui-classic/blob/master/app/controllers/application_controller.rb#L1326-L1336)