Rolling back changes to original MemoryStore
We did remove a number of monkey patches in #20610 f9e68c593a53
that added delete_sessions to each store.

But rolled back the implementation of MemoryStore.
The one based upon CacheStore had strange session_ids
so using the one based upon AbstractStore works better for now
Since it is only used for test, it makes the most sense.

Did remove the `private` declaration for the find/delete methods