Implement Target#column_type
Don't think we even Target.new, but want a base implementation in there
so we can use `&.column_type` over `try(:column_type)`

Target is an undetermined entity. We default to non-numeric/:string.
So even thought we could lookup the attribute type off of the target,
hard-coding to a string was more in line with the existing code.

will leverage column_type/numeric? in the next commit that converts
parse_field_or_tag => parse commit
They all change the same lines