Workaround dev issue with nuage needing qpid_proton
Nuage depends on qpid_proton which we can't require on mac

In dev, if we eager load the application, such as with `rake zeitwerk:check`, this file blows up
on mac because qpid_proton is an optional dependency but actively referenced in messaging_handler.rb.

Only ignore loading qpid_proton here on mac.