introduce MiqExpression#prune_exp
The goal is to prune nodes from an expression tree that are not applicable
to our target language

When generating sql, we only want to process nodes that are sql friendly.
In the case that both sql and ruby are in and OR clause, we will need to run this
in ruby, so both are removed. Logically you can not split up the entries in the OR clause.

When generating ruby, we only want to process ruby nodes.
Or in the case that both sql and ruby are needed together, we deal with both.

This is based upon the original preprocess_for_sql, but it allows flexibility
for choosing to keep ruby OR sql.

This all came about because in Rbac we are filtering in both sql and ruby.
The sql filter was reduced, but the ruby filter contained both the ruby and sql filters.
It doesn't make sense to run the same sql filters when processing ruby
since those have already been run.

In a followup commit, there is a note around demorgan's law which may help you
understand a little more about how these various components are splitup