Converge batch_size and concurrent_requests
Before

- concurrent_requests, thought it existed in core, was vmware only.
  It was split up into historical, hourly, and realtime.

- batch_size was per provider (but only implemented by vmware)
  It just has a single option.

After

- concurrent_requests defined in the provider (but only implemented by vmare)
  It is split up into historical, hourly, and realtime.

Reason for change

While  batch_size may be 250, concurrent_requests was then applied and we were sending
only 20 requests for realtime and 1 for historical.

So in the end, we only are sending 1 request to the provider. The optimization has basically
been disabled.

This gives better values for concurrent requests, and matches the requests sent
to the collector over the queue and from the collector to the provider

This requires changes in both vmware and core.

A number of specs had the settings for vmware when they were testing openstack.
This has been corrected.