Move Target#column_type from child implementations to parent.
Currently it doesn't implement `column_type`.
Adding it lets us transition from `parse_field_or_tag.try(:column_type)` to `&.column_type`.

For errors, we return `col_type=nil` (stored in `:data_type` for `get_col_info`)
So for that reason returning that value.

numeric? is no longer needed for tag, since the column type is :string, and that is not numeric

Had briefly thought Target#column_type == :string. There are a bunch of places of code that
override it, but using `get_col_info` as the definitive set of values to return for this column
in all cases.

Will leverage column_type and numeric? in the next commit that converts
parse_field_or_tag => parse
No need to change all the same lines.