Fix cloud provision quota memory and cpu counting
Cloud providers are not required to use flavors to set VM memory and cpu
allocation during provisioning (for example
IbmCloud::PowerVirtualServers::CloudManager). These 'number_of_cpus' and
'memory' methods are expected to return int values. In the case where a
'cloud' type provision is associated with a flavor without 'memory' or
'cpus' values the non-cloud logic is better than returning nil.
Depending on how the provider implements the provision request, the
return value will be something useful or just 0, avoiding raising an
exception.