Normalize values returned by Target subclasses
- Target#column_type == nil for unknown (in all cases)

A Field used to thrown an ArgumentError when the association is not known
But since no object would be returned from parse with a bad model - that would be nil.

Now, if the model or the associations are bad, it returns nil.

- Tag#valid? only if there is a target. (value used to be undefined)
- Target#includes are now {} when they are unknown

Documented when an error could be returned.
Did contemplate dropping these errors. (these are internal methods)