Convert parse_field_or_tag to Target.parse
All the try(:column_type) / &.column_type calls are converted to column_type.

There are a few cases where column_type is defaulted the type to string.
This is due to the non-pedantic nature of our expression.

These cases should be throwing errors, but we'll cross that bridge later.

MiqExpression#fields is using a kind_of?(). Sorry.
I was thinking about using valid? (which had a few more implications) or
introducing a different method, but this seemed the best way forward.

In the future, we want to store the field in the tree anyway
(possibly in expression["field_field"] and expression["value_field"])
so this code should go away soon enough anyway. (famous last words)