Introduce MiqExpression::Target#parse
Target#parse is a generic parse method. It handles any expression and returns
the appropriate type.

We were not taking advantage of parse!. The code is not pedantic and just
has a fallback. So passing an "invalid" object that has all the right values.

I did like the original rescue and the fallback in get_col_info.
If any method failed (typically bad associations) we would get the defaults for all of the attributes.
But the tests show that my concern there was not warranted. We know it is an all or nothing kind of
response, even without the rescue.

Since most (all?) of the calls into parse_field_or_tag were originally get_col_info,
following that method's return values as the defaults gets us to a better place.

The calling code is setup to handle a bad argument and not be too pedantic.
Many get_col_info calls have been converted across to use parse_filed_or_tag,
but they don't handle ArgumentErrors. Now that these can no longer be thrown for
bad associations removes those edge cases.