pass column_type as a symbols for quote*(val, column_type)
Most uses of column_type are a symbol or a nil:

- Target#column_type is a symbols or nil.
- options2hash[:type] is a string (the only case I can find that uses a string)
- hardcoded string/symbol
- Target#column_type || hard coded string/symbol

The `to_s` in `quote` can handle a string or a symbol, so even if we missed any
cases, we are still protected.
Added support for nil, which will work the same as if :string were passed in.
We changed all callers to pass a nil or a symbol.

The `to_s` in `quote_human` can also handle all three.
Added support for nil to behave like a string (so we don't have to || the type)
all quote_human calls are passed symbols

all quote_by calls are passed type as a symbol

In the future, we can remove the to_s in quote and quote_human

In the future, we will have the actual field value in the tree, so we
will no longer need Target.parse in so many places to determine the target/column_type.

In the future, we're dropping <value>, so much of this will be removed.