metrics tests: ignoring historical date splits and object batching
it is getting confusing determining which hosts or vms get batched together.
So batches are exploded into multiple line items
And timings are condensed into contiguous ranges.

For testing that multiple rows were inserted for a time range, the queue can be counted
For testing that batches were sent, something similar is done.

This makes it a little trickier to ensure that rollups were generated correctly.
That will need to be done directly with the queue.

Currently the queue_timings method is only used for historical tests. It would simplify
our tests and make them more consistent if this method were used across all tests