Fix normal priority reports not being run in reporting workers
Fixes #22562

With this change, for services with a dedicated queue_name and worker, such as reporting, high
priority work is run in the 'generic' queue with HIGH_PRIORITY so priority workers or even generic
workers can pick it up immediately. Anything but high priority work or higher is run in the
service's queue, such as 'reporting', and the dedicated worker will process this message following
priority processing.

Why? Reports are often started by a user who is actively waiting for the result so we need a way
to expedite their request in workers who deals with high priority work.  At the same time, if a
user isn't actively waiting for a report, it should be handled by reporting workers and not sit
behind other items in the generic queue as it's far easier to scale these workers up and down
as needed.