fix supports_feature class method test
The purpose of these tests are if the underscore method is defined.
Shouldn't test with supports? method

NOTE: 3 lines down, there is the supports?() version of the test

incorrectly added in  1623134c1c2bf

It may have been a remnant when we defined supports :feature and wanted supports_feature?
and when we defined supports_feature and wanted supports?(:feature)
(availability mixin)