Allow supports?(feature) to work without requiring supports_not
`supports :feature` helps us determine if a model has implemented a given feature.

Previously, a call to `supports? :feature` required that a `supports` or a `supports_not`
was previously declared on the model class or parent class.

Before
======

All models had `supports_not :feature` on the base of all trees.
We automatically called `supports_not` for the 4 crud methods.

If we forget to declare `supports_not :feature` for a model, or if the feature
is misspelled, the code will throw a runtime error. (good / bad)

After
=====

A model defaults to not supported whether a `supports_not` is used or not.

No runtime errors surprise us.
No typos show up in runtime either.

Long play
====

- Don't define `supports_not` crud methods (4 of them)
- Remove all `supports_not` on base classes (152 of them)
- Don't define any `supports_#{feature}?` methods and use a hash lookup
instead.