Remove User#unlock_queue use of put_or_update
`MiqQueue#put_or_update` is a powerful method but manipulates a queue in a way foreign to (all?) other implementations of queues.
We have been getting away from this method and are down to 3 uses.

When a user attempts too many logins, their account is locked. `User#unlock_queue` sets a task in the future to unlock the account after a timeout.

This functionality can be implemented using `MiqQueue#put_unless_exists` instead, getting us closer to the goal of dropping all usage of 'put_unless_exists`.

This method does introduce one change, and while I prefer the current behavior, the change is worth it.

Before
======

A failed login attempt resets the unlock timer.

Once a user attempts too many logins, their account is locked for a set amount of time.
Further attempts during the loc period further resets the timer.

After
=====

A failed login attempt does not reset the unlock timeout.

Once a user attempts too many logins, their account is locked for a set amount of time.
Further attempts fail, but that does not change the amount of time the account is locked.