vim_performance_state_for_ts - don't find(now)
since perf_capture_state does a find for now, not reason to try and find it first
have tried to remove this extra query a few times.

previous attempts deleted this line.
this one is different as we added the unless. there were a few cases that we do
want a find here. (luckily there are tests around this)