Fix migrations being run from a plugin
Before this commit, when running bin/setup on a fresh checkout of
ui-classic, the migrate_database step would fail because it would run in
the spec/manageiq directory as opposed to the plugin root directory.

What happens is that the spec/manageiq directory doesn't have a
Gemfile.lock, because the bundle install happens in the plugin root, and
so when migrate_database runs db:migrate from the spec/manageiq
directory it fails saying it can't find certain gems.

This commit changes it to run db:migrate from the plugin root directory.
Additionally, I've added a create_database step, which is a noop when
the database exists, but also simplifies brand new checkout runs.