Add methods for env vars in ManageIQ::Environment
Using boolean environment variables correctly can be tricky because the
values come in as strings rather than booleans in ruby so a simple `if
ENV["CI"]` will be true for `CI=true` AND `CI=false`.

Add methods that encapsulate this logic and present simple question
methods that return `true` or `false`.