Add deep delete functionality
This traverses an object tree, deleting any records in associations that have no ruby callbacks
It uses destroy on objects that have callbacks

The objects are loaded with embedded scopes, so N+1 for fetching those objects are removed.

This does not properly handle constraints.
That is why some PRs were put in to clean up incorrect constraints.
This will work without the constraints, but the warnings no longer show up.