Delete the container_manager reconnect_block
The reconnect_block for the container_manager was added in case
inventory was accidentally deleted via refresh the next refresh would
"bring it back".

While this is an interesting preemptive fix it would be better for us to
identify that there is an issue and fix it rather than trying to
partially fix it live.