Rails 7 removed clear_association_cache
In the event the association's objects are being changed outside
the in memory objects, we can do targetted reloads of the associations
to get similar results. Reloading the main object will blow up if the
object is not yet persisted and also discards any changes on that object.

In the case of ExtManagementSystem#clear_association_cache,
this was added when storages method wasn't a real association
so we manually built the ivar storages from the current hosts
so we needed to wrap the superclass clear_association_cache with
a clearing of our storages cache so it can be built clean.

With storages being an actual association on Ems, this is no longer
needed.

See:
commit dd5886d00a2d5f31ccf504c391aad93deb014eb8
Author: Ryuta Kamizono <kamipo@gmail.com>
Date:   Mon May 3 12:13:19 2021 +0900

    Remove unused `clear_association_cache`

    It is no longer used since #41112.