optimize supports_feature_mixin helper
This ensures unsupported_reason is stubbed for both supports and non-supports cases
(the number of calls to supports is reducing)

Since supports? and #unsupported_reason is just a minor shim over .unsupported_reason,
left those methods as is. Only stubbing .unsupported_reason

This also removes the allow_any_instance_of