Add a GenericMailer method that creates a MiqTask
There is currently a `GenericMailer.deliver_queue` which puts the deliver
request on the queue but no task is created for tracking the state/status
of the request.

This adds a `GenericMailer.deliver_task` method that wraps
`.deliver_queue` but creates a `MiqTask` and adds a queue callback to
the `MiqQueue` options.